Cache location reverse lookup results
As the openstreetmap reverse lookup has now become more strict in its
requests/responses, we'll now heavily cache results to the nearest 2
decimal places. As we only want the city and state, this will be
accurate enough to cache.

TODO: Clear location cache after a number of days/weeks/months